Fix contact without any relation
After migration contacts were without any relation
to member. Because the migration were written in 
ruby code and the belongs_to was written after 
migration without running them again with the 
backup from the production.
This commit searches a member based on the 
username from user by the name of the club_email. 
This migration does not effect any rollback 
action. If a rollback is triggered nothing will 
be change. This migration should be deleted after 
it finished because it will overwrite the member 
id on second run.