Additional player fields, fixes #56 (#118)
* Add personal moment and players ideal

* Add more links to adminstration panel

* Small cleanups

* Add n:m relation for players and teams

* Add simple overview to add players

* Add more translations

* Use reponder gem

* Change the mailing logic, fixes #44

Player form the club team are seen at first,
all active player with the possibility from the 
club team are automatically checked. All 
other player are available in the dropdown.
The list containing all names is removed and #44
is no issue anymore.

* Replace email with status of the player

* Clean up templates

* Update schema

* Split template into partials

* Fix migration

This fix is needed because the production database
already created the players_teams table.